finished remove linked lists(2)
