+++ .vscode/settings.json
@@ -2,10 +2,10 @@
     "python.testing.unittestArgs": [
         "-v",
         "-s",
-        "./Medium",
+        "./RemoveLinkedList(2)",
         "-p",
         "*test.py"
     ],
-    "python.testing.pytestEnabled": false,
-    "python.testing.unittestEnabled": true
+    "python.testing.unittestEnabled": true,
+    "python.testing.pytestEnabled": false
 }
