--- Medium/RemoveLinkedList(2)/remove_linked_list_elements.py
@@ -1,29 +1,0 @@
-from list_node import ListNode  # Import ListNode class (assuming it's in a separate file)
-
-class Solution:
-  def remove_nodes_with_count(self, head, target_counts):
-    """
-    Removes nodes from the linked list based on target values and desired removal counts.
-
-    Args:
-        head: The head node of the linked list.
-        target_counts: A list of tuples where each tuple represents a target value and its desired removal count.
-
-    Returns:
-        The modified head node of the linked list.
-    """
-
-    prev = None
-    cur = head
-
-    while cur != None:
-        prev = cur
-        cur = cur.next
-
-    return head
-  
-
-
-  
-  
-
