+++ RemoveLinkedList(2)/__pycache__/list_node.cpython-312.pyc
