+++ RemoveLinkedList(2)/__pycache__/remove_linked_list_elements_test.cpython-312-pytest-8.2.1.pyc
