+++ RemoveLinkedList(2)/list_node.py
